+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/impl/StringTableValueImpl.java
@@ -1,8 +1,8 @@
 package com.kaylerrenslow.armaDialogCreator.arma.stringtable.impl;
 
 import com.kaylerrenslow.armaDialogCreator.arma.stringtable.Language;
-import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableKey;
 import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableValue;
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableMap;
 import org.jetbrains.annotations.NotNull;
 
@@ -10,18 +10,20 @@ import org.jetbrains.annotations.NotNull;
  @author Kayler
  @since 12/12/2016 */
 public class StringTableValueImpl implements StringTableValue {
-	private final StringTableKey key;
+
 	private final ObservableMap<Language, String> values;
 
-	public StringTableValueImpl(@NotNull StringTableKey key, @NotNull ObservableMap<Language, String> values) {
-		this.key = key;
+	public StringTableValueImpl(@NotNull ObservableMap<Language, String> values) {
 		this.values = values;
 	}
 
+
 	@Override
 	@NotNull
-	public StringTableKey getKey() {
-		return key;
+	public StringTableValue deepCopy() {
+		StringTableValue copy = new StringTableValueImpl(FXCollections.emptyObservableMap());
+		copy.getLanguageTokenMap().putAll(this.getLanguageTokenMap());
+		return copy;
 	}
 
 	@NotNull